@@ -19,12 +19,12 @@
     },
     "minimum-stability": "stable",
     "require": {
-        "php": ">=7.4.0",
+        "php": ">=8.0",
         "yiisoft/yii2": "~2.0.45",
         "yiisoft/yii2-bootstrap5": "~2.0.2",
         "yiisoft/yii2-symfonymailer": "~2.0.3",
         "yiisoft/yii2-authclient": "*",
-      "ext-mbstring": "*"
+        "kreait/firebase-php": "^7.0"
     },
     "require-dev": {
         "yiisoft/yii2-debug": "~2.1.0",
