@@ -10,6 +10,7 @@ use yii\filters\AccessControl;
 use yii\web\Controller;
 use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
+use yii\web\Response;
 use yii\web\ServerErrorHttpException;
 
 class MyController extends Controller
@@ -31,39 +32,54 @@ class MyController extends Controller
         ];
     }
 
-    public function actionIndex()
+    /**
+     * Страница просмотра личных объявлений
+     *
+     * @return string - код страницы
+     */
+    public function actionIndex(): string
     {
         $id = Yii::$app->user->id;
         $myTicketsProvider = new ActiveDataProvider([
-            'query' => Ticket::queryMyTickets($id)
+            'query' => Ticket::getMyTickets($id)
         ]);
         return $this->render('index', [
             'myTicketsProvider' => $myTicketsProvider
         ]);
     }
 
-    public function actionComments()
+    /**
+     * Страница просмотра личных объявлений с комментариями
+     *
+     * @return string - код страницы
+     */
+    public function actionComments(): string
     {
         $userId = Yii::$app->user->id;
         $ticketProvider = new ActiveDataProvider([
-            'query' => Ticket::queryMyTicketsWithComments($userId)
+            'query' => Ticket::getMyTicketsWithComments($userId)
         ]);
         return $this->render('comments', [
             'ticketProvider' => $ticketProvider
         ]);
     }
 
-
     /**
+     * Удаление выбранного объявления
+     *
+     * @param int $id - id объявления
+     * @return Response - код страницы
      * @throws ForbiddenHttpException
      * @throws ServerErrorHttpException
      * @throws NotFoundHttpException
      */
-    public function actionDelete($id){
+    public function actionDelete(int $id): Response
+    {
         $ticket = Ticket::findOne($id);
-        if (!$ticket){
+        if (!$ticket || $ticket->status === 0){
             throw new NotFoundHttpException ('Объявление не найдено');
         }
+
         if (Yii::$app->user->can('editAllTickets', ['author_id' => $ticket->user_id])) {
             if ($ticket->deleteTicket()) {
                 return $this->redirect('/my');
@@ -75,17 +91,22 @@ class MyController extends Controller
         }
     }
 
-
     /**
+     * Удаление выбранного комментария
+     *
+     * @param int $id - id комментария
+     * @return Response - код страницы
      * @throws ForbiddenHttpException
      * @throws ServerErrorHttpException
      * @throws NotFoundHttpException
      */
-    public function actionCommentout($id){
+    public function actionCommentout(int $id): Response
+    {
         $comment = Comment::findOne($id);
-        if (!$comment){
+        if (!$comment || $comment->status === 0){
             throw new NotFoundHttpException ('Объявление не найдено');
         }
+
         if (Yii::$app->user->can('editAllTickets', ['author_id' => $comment->ticket->user_id])) {
             if ($comment->deleteComment()) {
                 return $this->redirect('/my/comments');
