@@ -113,57 +113,10 @@ class Ticket extends ActiveRecord
         return $this->hasOne(User::class, ['id' => 'user_id']);
     }
 
-    public static function queryFreshTickets(): ActiveQuery
-    {
-        return Ticket::find()->where(['status' => 1])->orderBy('date_add DESC');
-    }
-
-    public static function queryPopularTickets(): ActiveQuery
-    {
-        return
-            Ticket::find()
-                ->join('LEFT JOIN', 'comment', 'comment.ticket_id = ticket.id')
-                ->groupBy('ticket.id')
-                ->having('COUNT(comment.id) > 0 AND status = 1')
-                ->orderBy('COUNT(comment.id) DESC');
-    }
-
-    public static function queryTicketsInCategory($categoryId): ActiveQuery
-    {
-        return
-            Ticket::find()
-                ->select('id, status, header, photo, price, type, text, ticket_category.category_id as category_id')
-                ->leftJoin('ticket_category', 'ticket_category.ticket_id = ticket.id')
-                ->having('ticket.status = 1 and ticket_category.category_id = ' . $categoryId)
-                ->orderBy('ticket.id DESC');
-    }
-
-    public static function queryMyTicketsWithComments($userId): ActiveQuery
-    {
-        return
-            Ticket::find()
-                ->leftJoin('comment', 'comment.ticket_id = ticket.id')
-                ->where([
-                    'ticket.user_id' => $userId,
-                    'ticket.status' => 1,
-                    'comment.status' => 1])
-                ->groupBy('ticket.id')
-                ->having('COUNT(comment.id) > 0')
-                ->orderBy('MAX(comment.id) DESC');
-    }
-
-    public static function queryMyTickets($userId): ActiveQuery
-    {
-        return
-            Ticket::find()
-                ->where([
-                    'user_id' => $userId,
-                    'status' => 1
-                ])
-                ->orderBy('id DESC');
-    }
-
     /**
+     * Метод удаления объявления (скрываем на сайте, но в базе оставляем)
+     *
+     * @return bool
      * @throws ServerErrorHttpException
      */
     public function deleteTicket(): bool
@@ -180,9 +133,86 @@ class Ticket extends ActiveRecord
             $transaction->commit();
         } catch (Exception $e){
             $transaction->rollback();
-            throw new ServerErrorHttpException('Проблема на сервере. Объявление удалить не удалилось.');
+            throw new ServerErrorHttpException('Проблема на сервере. Объявление удалить не удалось.');
         }
         return true;
     }
 
+    /**
+     * Метод получения объявлений (в начале - с макисмальным id объявления)
+     *
+     * @return ActiveQuery объявления, в пордяке от самого нового
+     */
+    public static function getFreshTickets(): ActiveQuery
+    {
+        return Ticket::find()->where(['status' => 1])->orderBy('date_add DESC');
+    }
+
+    /**
+     * Метод получения объявлений (в начале - с макисмальным количеством комменатриев)
+     *
+     * @return ActiveQuery объявления, у которых есть хотя бы 1 комментарий
+     */
+    public static function getPopularTickets(): ActiveQuery
+    {
+        return
+            Ticket::find()
+                ->join('LEFT JOIN', 'comment', 'comment.ticket_id = ticket.id')
+                ->groupBy('ticket.id')
+                ->having('COUNT(comment.id) > 0 AND status = 1')
+                ->orderBy('COUNT(comment.id) DESC');
+    }
+
+    /**
+     * Метод получения объявлений заданной категории
+     *
+     * @param int $categoryId - id категории
+     * @return ActiveQuery объявления, в пордяке от самого нового
+     */
+    public static function getTicketsInCategory(int $categoryId): ActiveQuery
+    {
+        return
+            Ticket::find()
+                ->select('id, status, header, photo, price, type, text, ticket_category.category_id as category_id')
+                ->leftJoin('ticket_category', 'ticket_category.ticket_id = ticket.id')
+                ->having('ticket.status = 1 and ticket_category.category_id = ' . $categoryId)
+                ->orderBy('ticket.id DESC');
+    }
+
+    /**
+     * Метод получения объявлений у которых есть комменатрии, автором которых является текущий юзер
+     *
+     * @param int $userId - id юзера
+     * @return ActiveQuery объявления, в пордяке от объявления, с самым свежим комментарием
+     */
+    public static function getMyTicketsWithComments(int $userId): ActiveQuery
+    {
+        return
+            Ticket::find()
+                ->leftJoin('comment', 'comment.ticket_id = ticket.id')
+                ->where([
+                    'ticket.user_id' => $userId,
+                    'ticket.status' => 1,
+                    'comment.status' => 1])
+                ->groupBy('ticket.id')
+                ->having('COUNT(comment.id) > 0')
+                ->orderBy('MAX(comment.id) DESC');
+    }
+
+    /**
+     * Метод получения объявлений, автором которых является текущий юзер
+     *
+     * @param int $userId - id юзера
+     * @return ActiveQuery объявления, в пордяке от объявления, с самым свежим комментарием
+     */
+    public static function getMyTickets(int $userId): ActiveQuery
+    {
+        return
+            Ticket::find()
+                ->where([
+                    'user_id' => $userId,
+                    'status' => 1
+                ])
+                ->orderBy('id DESC');
+    }
 }
