@@ -7,6 +7,11 @@ class TicketHandler
     const TICKET_BUY = 'buy';
     const TICKET_SELL = 'sell';
 
+    /**
+     * Карта типов объявлений
+     *
+     * @return array
+     */
     public static function getTypeMap(): array
     {
         return [
@@ -15,8 +20,15 @@ class TicketHandler
         ];
     }
 
-    public static function getLabel($type){
-        if ($type === 'buy') {
+    /**
+     * Метод полуения типа объвления на русском языке
+     *
+     * @param string $type тип объявления
+     * @return string
+     */
+    public static function getLabel(string $type): string
+    {
+        if ($type === self::TICKET_BUY) {
             return 'Куплю';
         } else {
             return 'Продам';
