@@ -6,11 +6,15 @@ use yii\web\ServerErrorHttpException;
 
 class UploadFile
 {
-
     /**
+     * Метод загрузки файла в заданную директорию
+     *
+     * @param $fileForSave - загружаемый файл
+     * @param string $direction - директория
+     * @return string
      * @throws ServerErrorHttpException
      */
-    public static function upload($fileForSave, $direction): string
+    public static function upload($fileForSave, string $direction): string
     {
         $name = uniqid() . '.' . $fileForSave->getExtension();
         if ($fileForSave->saveAs("@webroot/uploads/{$direction}/" . $name)) {
