@@ -20,7 +20,7 @@
     var userChatField = chat.querySelector('[name="chat-message"]');
 
     document.addEventListener('click', function (evt) {
-      evt.preventDefault();
+      //evt.preventDefault();
       if (evt.target === buttonToOpenChat) {
         evt.preventDefault();
         openChat();
