@@ -3,10 +3,11 @@
 /** @var Comment $comment */
 
 use app\models\Comment;
+use yii\helpers\Html;
 use yii\helpers\Url;
 ?>
 
-<li>
+<li <?=(Yii::$app->request->url === '/my/comments') ? 'js-card' : '' ?>>
     <div class="comment-card">
         <div class="comment-card__header">
             <a href="#" class="comment-card__avatar avatar">
@@ -17,5 +18,8 @@ use yii\helpers\Url;
         <div class="comment-card__content">
             <p><?=$comment->text; ?></p>
         </div>
+        <?php if (Yii::$app->request->url === '/my/comments'): ?>
+            <?=Html::a('Удалить', Url::to(['my/commentout/'. $comment->id]), ['class'=>'comment-card__delete js-delete']); ?>
+        <?php endif; ?>
     </div>
 </li>
